panfrost/drm-shim: go through uintptr_t when casting a query pointer
The panthor dev_query handlers cast the uAPI's __u64 pointer field
straight to a struct pointer.  On a 32-bit host that is a cast to a
pointer of different size, and with -Werror it breaks the build --
which starts to matter once the arm32 CI build compiles drm-shim:

  panfrost_noop.c:168:10: error: cast to pointer from integer of
  different size [-Werror=int-to-pointer-cast]

and the same at the csif, timestamp and group priorities queries.

Cast through uintptr_t, as the DRM uAPI user pointer convention wants.

Fixes: 76397f76c944 ("panfrost: Add support for Panthor in drm-shim")
Fixes: 6a6a8c6cffa9 ("panfrost: add support for DRM_PANTHOR_DEV_QUERY_TIMESTAMP_INFO on drm-shim")
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/43489>

+++ src/panfrost/drm-shim/panfrost_noop.c
@@ -165,7 +165,7 @@ panthor_ioctl_dev_query(int fd, unsigned long request, void *arg)
    switch (dev_query->type) {
    case DRM_PANTHOR_DEV_QUERY_GPU_INFO: {
       struct drm_panthor_gpu_info *gpu_info =
-         (struct drm_panthor_gpu_info *)dev_query->pointer;
+         (struct drm_panthor_gpu_info *)(uintptr_t)dev_query->pointer;
 
       gpu_info->gpu_id = pan_get_gpu_id() << 16;
       gpu_info->gpu_rev = 0;
@@ -194,7 +194,7 @@ panthor_ioctl_dev_query(int fd, unsigned long request, void *arg)
    }
    case DRM_PANTHOR_DEV_QUERY_CSIF_INFO: {
       struct drm_panthor_csif_info *csif_info =
-         (struct drm_panthor_csif_info *)dev_query->pointer;
+         (struct drm_panthor_csif_info *)(uintptr_t)dev_query->pointer;
 
       unsigned arch = pan_get_gpu_id() >> 12;
 
@@ -207,7 +207,7 @@ panthor_ioctl_dev_query(int fd, unsigned long request, void *arg)
    }
    case DRM_PANTHOR_DEV_QUERY_TIMESTAMP_INFO: {
       struct drm_panthor_timestamp_info *timestamp_info =
-         (struct drm_panthor_timestamp_info *)dev_query->pointer;
+         (struct drm_panthor_timestamp_info *)(uintptr_t)dev_query->pointer;
 
       /* Noop values */
       timestamp_info->timestamp_frequency = 0;
@@ -218,7 +218,7 @@ panthor_ioctl_dev_query(int fd, unsigned long request, void *arg)
    }
    case DRM_PANTHOR_DEV_QUERY_GROUP_PRIORITIES_INFO: {
       struct drm_panthor_group_priorities_info *priorities_info =
-         (struct drm_panthor_group_priorities_info *)dev_query->pointer;
+         (struct drm_panthor_group_priorities_info *)(uintptr_t)dev_query->pointer;
 
       /* Default values */
       priorities_info->allowed_mask =
